requirements.txt
annotated-types==0.7.0
anthropic==0.72.0
anyio==4.5.2
certifi==2025.11.12
charset-normalizer==3.4.4
distro==1.9.0
docstring-parser==0.17.0
exceptiongroup==1.3.1
faiss-cpu==1.8.0.post1
filelock==3.16.1
fsspec==2025.3.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
jinja2==3.1.6
jiter==0.9.1
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==25.0
pandas==2.0.3
pillow==10.4.0
pydantic==2.10.6
pydantic-core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.17.0
sniffio==1.3.1
sympy==1.13.3
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.2.2
tqdm==4.67.1
transformers==4.46.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
